requirements.txt
requests
pandas
numpy
streamlit
plotly
python-dotenv
pytest
